+++ src/game/tests.rs
@@ -494,6 +494,7 @@ mod vintage_cube_memory_lapse;
 mod vintage_cube_metamorphosis_fanatic;
 mod vintage_cube_mightform_harmonizer;
 mod vintage_cube_mine_collapse;
+mod vintage_cube_minsc_and_boo;
 mod vintage_cube_mishras_bauble;
 mod vintage_cube_monarch;
 mod vintage_cube_monstrous_rage;